Attempt to fix parsing of polynomials with constant coefficients where the coeffs is a rational function field.
The function used for parsing polynomials (p_Read) stopped if the first
variable encountered was not in the ring. This doesn't work when the
coefficient ring is a transcendental extension field, since p_Read is
called recursively but the ring it receives as an argument doesn't know
about the variable names of outer rings.

We return the value parsed so far instead of cancelling the operation.
This seems to fix the problem, hopefully without breaking much else. :)