Modified the omalloc configuration for factory/ and made toplevel use it
The factory/ configure now has four options for omalloc:
 --enable-omalloc
 --with-omalloc-dir=<path> (to point to omalloc's main install)
 --with-omalloc-extra-dir=<path> (in case there needs a second
                                  -- think out-of-tree builds)
 --enable-omalloc-trust (to trust omalloc will be found at build-time
                         but don't check at configure-time)

That last is especially useful to build factory/ inside singular,
because in that case, omalloc.h includes omTables.h... which will exist
only during the build!

Finally, the toplevel Makefile.am and configure.ac are modified, so
that the correct options are passed to factory/configure.