libfac: don't rebuild object files all the time.
libfac/Makefile.in puts object files into the OPTOBJ directory,
which it creates by some target and has it as prerequisite of
all the files generated in this directory.
This has the side effect of make considering every of those files
out of date once a any other file is created in that directory.
This means that when running make multiple times, those files will
be recreated every time.
To avoid that replace the OPTOBJ dependency with a
order-only-dependency (behind '|'), only requesting the directory
to be created but not causing the files to get outdated.