chg: name changes to be mathematically correct
some names were not mathematically correct but were like that just
because I copied them from gfanlib one to one.
however since I plan on developing a convex package that is a mixture
between gfanlib and polymake, and since the polymake names were chosen
with more mathematical care, I decided to make a few namechanges.
also this was an issue which has come up in previous singular meetings