chg: various changes, completed documentation
removed makePolytope
an extra function to convert a cone into a polytope is unnecessary,
will write a cast for that in the future

removed scalePolytope
an extra function to scale a polytope is unnecessary,
since one can just write p*k, p polytope, k integer.

removed isComplete
this is not working as expected (gfanlib problems),
should user need such a functionality,
he/she needs to resort to the polymake interface.

changed insertCone/removeCone
input is checked for compatibility by default,
should the user wish to circumvent this check (which is slow)
the user needs to pass an extra int argument which is 0.

removed various chunks of code that were commented out