Add const qualifier to some methods
Note that Q_poly::is_equal() was not made const -- it reduces the
polynomial, which of course modifies it. This seems odd, as it is
different from the behaviour of the two other polynomial classes. It
also seems odd to have a method name is_equal() with side effects.
Perhaps this is a bug?