chg: update results for Manual/assPrimes.tst
assPrimes() calls modStd(), and modStd() uses random(), but the number
of times that random() is called by modStd() has changed. Thus after
calling modStd(), assPrimes() uses a different ('random') coordiante
change, and therefore the output of this test has changed. But the
ideals in the result are actually the same (modulo order). Note that
due to additional GB computations, the test now takes more time.