autotools chain: decent generated source tarball
In the script `autogen.sh', the macro directory m4 is no more considered
as an include directory in order to include in the generated source tarball
the needed m4 macro files.

`doc/Makefile.am' needed material to generate the html documentation is added
to EXTRA_DIST; not needed material is removed from the list.