Allow for arbitrary number of generators
Gives some variables more telling names.

Slight restructuring of rootIsolation to allow for arbitrary number of
ideal generators:
Under certain conditions some reduced Groebner basis w.r.t. a lex
ordering has nvars(basering) generators, so pick that.
Otherwise we compute a triangular decomposition and apply rootIsolation
to all the ideals. This is done in rootIsolationTriangL with a check if
the found boxes overlap.

As rootIsolation does not depend on the proper amount of generators
anymore we solve the problem of boxes with no interior arising in
computations by projecting into a subring.

After these changes the first list returned by rootIsolation should
always be empty if eps == 0.