Remove some useless files
The files AUTHORS, ChangeLog and NEWS were empty, let's remove them
(they probably were there because automake complains if they are
missing -- unless one passes the `foreign` option to it, which we do).

The file clone_autogen_generated_stuff.sh clones a 6 year old repository,
apparently in an effort to not make the user need autotools installed.
Which seems like a bad idea to start with, but it also hasn't been kept
up-to-date. So let's just ditch it.