Tst/regress.cmd: avoid multiples of 256 as exit code
Running this script gave me a test failure, yet the exit code was 0. Digging
closer, it turns out the perl script tried to use exit code 256 for some
reason I don't understand. But exit codes are a single byte, so any multiple
of 256 gets mapped to 0, giving the false impression that the test suite
passes when it actually is failing.