Add no_rebase to libpolys's Makefile.am
The restructuring of libpolys had the effect that all the header files were
installed into a flat directory, instead of conserving the directory structure.
This makes no difference to compiling the Singular binary, but when trying to
link against the libsingular.la library, this lead to the unfortunate problem,
that, for instance,

    #include <misc/auxiliary.h>

would fail, as auxiliary.h was never installed in a directoy named misc.