**/*: rename singular.hlp to singular.info.
The help file presently named "singular.hlp" is intended to be read
with GNU Info, which looks for an extension of ".info" by
default. Thus, at the moment, one must run "info singular.hlp" to read
the installed file, whereas "info singular" is more natural.

This commit replaces "singular.hlp" with "singular.info" throughout
the source tree so that the file ultimately has the name that GNU Info
or Emacs expect.

Closes: https://github.com/Singular/Singular/issues/1107