*/*.{c,h}: fix invalid function prototypes.
The C99 standard does not allow function prototypes (or inline
definitions) to have an empty argument list, such as

  int foo(); /* invalid */

This is caught by -Werror=strict-prototypes, which is likely to become
the default in later versions of GCC and clang. This commit corrects
all(?) of them by adding a "void" parameter.