omalloc/omAllocSystem.c: add another include for malloc.h.
There are two branches in this file, one for HAVE_MALLOC_SIZE and one
for HAVE_MALLOC_USABLE_SIZE. The former includes malloc.h, and the
latter needs to too; otherwise, malloc_usable_size() is undefined.

This is caught by -Werror=implicit-function-declaration, which is
likely to be enabled by default in future versions of GCC and clang.