fix broken module machinery in libpolys
The libpolys machinery to build modules appears broken.
This is because a  mixture of linker options and libtools options
is passed to a m4-macro (AX_APPEND_LINK_FLAGS) that filters only
linker options. This patch attempts to fix this issue by mimicking
the machinery currently used to build modules in the Singular folder.